requirements.txt
Flask
psycopg2-binary
python-dotenv
joblib
lightgbm
scikit-learn
tqdm